Move code to service
