+++ app/services/__init__.py
@@ -1,11 +1,11 @@
 from .events.role_manager import EventRoleManager
-from .events.timetable_contructor import EventTimetableConstructor
-from .events.shopping_contructor import EventShoppingConstructor
+from .events.timetable_constructor import EventTimetableConstructor
+from .events.shopping_list_constructor import ShoppingListConstructor
 from .tips.approver import TipApprover
 
 __all__ = [
     "EventRoleManager",
     "EventTimetableConstructor",
-    "EventShoppingConstructor",
+    "ShoppingListConstructor",
     "TipApprover",
 ]
