--- app/services/events/shopping_constructor.py
@@ -1,2 +1,0 @@
-class EventShoppingContructor:
-    pass
