+++ app/services/events/timetable_constructor.py
